refactor: create nav package to make uniform
