+++ features/id-check-wrapper/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/idcheckwrapper/internalapi/nav/IdCheckWrapperDestinations.kt
@@ -1,6 +1,7 @@
-package uk.gov.onelogin.criorchestrator.features.idcheckwrapper.internalapi
+package uk.gov.onelogin.criorchestrator.features.idcheckwrapper.internalapi.nav
 
 import kotlinx.serialization.Serializable
+import uk.gov.onelogin.criorchestrator.features.idcheckwrapper.internalapi.DocumentVariety
 import uk.gov.onelogin.criorchestrator.features.resume.internalapi.nav.ProveYourIdentityDestinations
 
 sealed interface IdCheckWrapperDestinations : ProveYourIdentityDestinations {
